Better way to stop ext extend
Use dynamic functions to stop the ext extend function so we don't have to use the untyped js
call so we don't have to work out the name space.

Can easily turn this into a macro to automatically move the mixins from the class to the init
function and other properties used in preprocessors as needed.